Сверстан экран коллекции пользователя
