+++ FakeNFT/Docs/statistics_decomposition.md
@@ -11,9 +11,9 @@
 Модуль №2:
 Верстка экрана профиля пользователя ( 2 часа, факт 4 часа)
 Содание перехода из рейтинга в профиль (1 час)
-WebView "перехода на страницу пользователя (1 час)
-Создание перехода на WebView (1 час)
-Верстка экрана коллекции пользователя (2 часа)
+WebView "перехода на страницу пользователя (1 час, факт 30 минут)
+Создание перехода на WebView (1 час , факт 30 минут)
+Верстка экрана коллекции пользователя (2 часа, 2 часа)
 Создание перехода на экран коллекции (1 час)
 ИТОГО в модуле: 8 часов.
