+++ FakeNFT/Scenes /Statistics/StatisticsModels/StatistiscsNFTModel.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+struct StatistiscsNFTModel {
+    let image: UIImage
+    let name: String
+    let price: Float
+    let rating: Int
+    let id: String
+    let isLike: Bool
+    let isInCart: Bool
+}
