+++ FakeNFT/Scenes /TabBarController/TabBarController.swift
@@ -31,7 +31,7 @@ final class TabBarController: UITabBarController {
         let catalogController = CatalogViewController()
         let cartController = CartViewController()
         // TODO:  Вернуть обратно перед ревью let statisticsController = StatisticsViewController()
-        let statisticsController = StatisticsProfileViewController()
+        let statisticsController = StatisticsCollectionViewController()
         profileController.tabBarItem = profileTabBarItem
         catalogController.tabBarItem = catalogTabBarItem
         cartController.tabBarItem = cartTabBarItem
